requirements.txt
streamlit
pandas
altair
numpy
qiskit
qiskit-ibm-runtime
ibm-watsonx-ai
python-dotenv
psutil
PyYAML
langchain-ollama
langgraph
scikit-learn
structlog
networkx
qdrant-client
httpx
cryptography
torch
torchvision
torchaudio
sentence-transformers
transformers
qiskit-aer
scipy
